+++ src/version.py
@@ -1,4 +1,4 @@
 
 APP_TITLE = "PlaySK Piano Roll Reader"
-APP_VERSION = "3.0"
+APP_VERSION = "3.1"
 COPY_RIGHT = "(C)Sasaki Katsumasa 2014-2023"
