+++ src/version_info.txt
@@ -4,8 +4,8 @@ VSVersionInfo(
   ffi=FixedFileInfo(
     # filevers and prodvers should be always a tuple with four items: (1, 2, 3, 4)
     # Set not needed items to zero 0.
-    filevers=(3, 0, 0, 0),
-    prodvers=(3, 0, 0, 0),
+    filevers=(3, 1, 0, 0),
+    prodvers=(3, 1, 0, 0),
     # Contains a bitmask that specifies the valid bits 'flags'r
     mask=0x3f,
     # Contains a bitmask that specifies the Boolean attributes of the file.
@@ -28,10 +28,10 @@ VSVersionInfo(
       StringTable(
         u'040904B0',
         [StringStruct(u'FileDescription', u'PlaySK Piano Roll Reader'),
-        StringStruct(u'FileVersion', u'3.0.0.0'),
+        StringStruct(u'FileVersion', u'3.1.0.0'),
         StringStruct(u'LegalCopyright', u'(C)Sasaki Katsumasa 2014-2023'),
         StringStruct(u'ProductName', u'PlaySK Piano Roll Reader'),
-        StringStruct(u'ProductVersion', u'3.0.0.0')])
+        StringStruct(u'ProductVersion', u'3.1.0.0')])
       ]), 
     VarFileInfo([VarStruct(u'Translation', [1033, 1200])])
   ]
